Crate rust_dsa

Structs

AnyStack

A stack that can hold arbitrary types.

BiMap

A bidirectional map.

BinaryHeap

A priority queue implementation backed by a binary heap.

BinomialHeap

A priority queue implementation backed by a binomial heap.

BloomFilter

A Bloom filter.

BumpAlloc

A bump allocator for type T.

Deque

A deque implementation backed by a ring buffer.

DiGraph

A weighted, directed graph.

FenwickTree

A Fenwick tree specialized for primitive integers and addition.

FibonacciHeap

A priority queue implementation backed by a Fibonacci heap.

GenericFenwickTree

A Fenwick tree that is generic over any type T and associative function F.

GenericTrie

A trie for mapping sequences of keys to values.

Graph

A weighted graph.

ImmutableVector

An immutable vector implementation with efficient edits/clones.

IntervalHeap

An interval heap implementation for use as a double-ended priority queue.

MedianHeap

A data structure for efficiently calculating a running median.

MinQueue

A FIFO queue that supports O(1) push, pop and find-minimum.

MinStack

A LIFO stack that supports O(1) push, pop and find-minimum.

Trie

A trie for mapping strs to values.

UnionFind

A disjoint-set data structure backed by a disjoint set forest.

VList

A dynamic array implementation backed by a VList.

Enums

Median

An enum representing the median in a MedianHeap.

Removed

An enum representing the items removed from a BiMap.

Functions

edit_distance

Returns the Levenshtein distance between two slices.

find_cycle

Returns a cycle in the graph, if one exists.

graham_scan

Returns a subset of the input points representing their convex hull.

heapsort

An implementation of the heapsort algorithm.

insertion_sort

An implementation of the insertion sort algorithm.

is_topological_sort

Returns true if the ordering is a topological sort for the graph.

majority_element

Returns the majority element of an iterator, if one exists.

mergesort

An implementation of the mergesort algorithm.

partition

Partitions the slice around the element at pivot_index.

quicksort

An implementation of the quicksort algorithm.

radix_sort

An implementation of the radix sort algorithm.

select

Returns a reference to the kth smallest element in the slice.

select_in_place

Returns a reference to the kth smallest element in the slice.

str_distance

Returns the Levenshtein distance between two strs.

topological_sort

Returns a topological sort of the graph, if one exists.